feat: flake refactor pt3
FINALLY working on the work mac. gonna add nix-darwin at some point, but for now
this works. had to change the home directory cuz that FUCKED me for a little bit
but got it working correctly. also made some changes to the actual root
configuration of the home manager stuff. all in all, this works super well for
my work stuff and is ready for nixos and nix-darwin methinks

Signed-off-by: Matt Provost <mprovost@cloudflare.com>